requirements.txt
urllib3
loguru
requests
pycryptodome
